Created component MentionUserTextArea.
